前台界面

+++ application/admin/controller/System.php
@@ -9,6 +9,13 @@ class System extends Base
 {
     protected $pageNum = "";
 
+    protected $pathName = 
+    [
+        'type6' => '用药关键字设置',
+        'other' => '其他设置',
+
+    ];
+
     public function _initialize()
     {
         parent::_initialize();
@@ -19,28 +26,37 @@ class System extends Base
      *
      * @return \think\Response
      */
-    public function index()
+    public function index($type = null)
     {
         $path = $this->request->path();
+        // dump($path);
+        // die();
         if(!$this->auth->check($path,session('admin_id')))
             return "<div style='margin-top:100px;text-align:center;'><h1> :(&nbsp;&nbsp;&nbsp;&nbsp;you have no permission!<h1></div>";
 
+        if($type){
+            $list = $this->model->where('type',$type)->select();
+            $this->assign('list', $list);
+    
+            $count = $this->model->where('type',$type)->count();
+            $this->assign('count',$count);
+
+            $this->assign('pathName',$this->pathName[$type]);
+            $this->assign('type',$type);
+            return $this->view->fetch();
+        }
+        
         $list = $this->model->where('type','site')->select();
         $this->assign('list', $list);
 
-        $list = $this->model->where('type','other')->select();
-        $this->assign('list1', $list);
-
-        $count = $this->model->count();
+        $count = $this->model->where('type','site')->count();
         $this->assign('count',$count);
+
+        $this->assign('type','site');
+        $this->assign('pathName','基本设置');
         return $this->view->fetch();
     }
 
-    /**
-     * 显示创建资源表单页.
-     *
-     * @return \think\Response
-     */
     public function add()
     {
         if($this->request->isPost())
@@ -71,9 +87,15 @@ class System extends Base
         if($this->request->isPost())
         {
             $param = $this->request->param();
+
+            if($param['value'] == $list->value and $param['description']==$list->description)
+                return json('您未做修改！');
             $res = $this->model->allowField(['value','description'])->save($_POST, ['id' => $param['id']]);;    
             if($res)
+            {
+                $this->clearCache();
                 return json(true);
+            }
             else
                 return json(false);
 
@@ -81,16 +103,12 @@ class System extends Base
         return $this->view->fetch();
     }
 
-    /**
-     * 删除指定资源
-     *
-     * @param  int  $id
-     * @return \think\Response
-     */
     public function delete($id =null)
     {
         if($this->request->isPost())
         {
+            if($this->model->where('id',$id)->column('type')['0'] == 'site')
+                return json('不允许删除');
             $res = $this->model->destroy($id);
             if($res)
                 return json(true);
@@ -104,6 +122,7 @@ class System extends Base
      * 删除缓存
      *
      * @param int $type  0:删除cache文件夹内容  1:删除temp文件夹内容  2:删除log文件夹内容  3:删除runtime文件夹内容
+     * //todo 1 2 3
      * @return void
      */
     public function clearCache($type = 0)
